Migracion_EFC/requirements.txt
# Dependencias principales del proyecto Migración EFC
# Generado automáticamente basado en análisis del código

# Base de datos y ORM
SQLAlchemy>=1.4.0,<2.0.0
pyodbc>=4.0.30

# Cliente HTTP
aiohttp>=3.8.0
requests>=2.28.0

# Configuración y variables de entorno
python-dotenv>=0.19.0

# Utilidades de fecha y tiempo (incluido en Python standard library)
# datetime, timedelta están incluidos

# Concurrencia (incluido en Python standard library)
# concurrent.futures, threading, asyncio están incluidos

# Parseo de argumentos (incluido en Python standard library)
# argparse está incluido

# Tipos de datos (incluido en Python standard library)
# dataclasses está incluido

# Dependencias opcionales para desarrollo y testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Para logging estructurado (opcional)
structlog>=22.0.0

# Para validación de datos (opcional)
pydantic>=1.10.0

# Para monitoreo (opcional)
prometheus-client>=0.15.0
